显示的当前位置的角度保留两位，修改了函数的命名，采集卡上传时出现故障的处理

+++ informationleft.h
@@ -16,7 +16,7 @@ public:
     ~informationleft();
 
 public slots:
-	void set_currentAngle(quint16 a);		//当前角度
+	void set_currentAngle(float a);			//当前角度
 	void set_groupcnt(quint16 a);			//采集组数分数？
 	void set_groupNum(quint32 a);			//采集组数百分数
 	void set_filename1(QString str);		//数据文件名1
